PUX-2581 Android UI Customisation (#30)
This adds support for customising the UI of the SDK on Android devices.
Themes allow customising the colors and fonts used in the UI. Colors
require a hexcode and supports both `RGB` and `ARGB`. Fonts support both
`.ttf` and `.otf` formats and must be placed in the
`android/app/src/main/res/font` folder. The file must be named in snake
case to be recognised. Then you simply use the name of the file (without
the extension) in your theme object.

```
const androidTheme =  {
  lightColors: {
    primary: "#FF32a852",
    background: "#888888",
    surface: "#888888",
    error: "#000000"
  },
  darkColors: {
    primary: "#888888",
    background: "#000000",
    error: "#cccccc"
  },
  typography: {
    bodyLarge: {
      font: "rainbow_2000"
    }
  }
}

const theme =  {
  android: androidTheme
}
```

That theme is then provided to the configure method.
```
TrueLayerPaymentsSDKWrapper.configure(Environment.Sandbox, theme).then(
  () => {
    log('Configure success')
  },
  reason => {
    log('Configure failed ' + reason)
  },
)
```

The theme is then passed to the native SDK via an intent.